Using red black tree to implement phonebook.h and phonebook.c
I use RBT to implement append() and findName() in phonebook.c to find lastName
in a dictionary. And I also rewrite the original entry struct to a smaller struct
by moving the useless items to a new struct.

In main.c I use words.txt as dictionary and "zyxel" as input. The timer of findName()
changed from sec to nsec.

In Makefile I add pstat to measure the performance of code and cache to clear the cache.