Use djb2 as hash function, modify time unit to msec.
Previous hash function was a check sum hash function, not good enough for string
hashing, since the djb2 hash function performs better. And the hash size 349903 is
a prime number just larger than 349900 which is the number of words in dictionary/words.txt.

I change diff_time unit to millisecond.